Fix incorrect plus(Long,...) behavior in js-joda
When 0L was passed as JS number argument, it failed to equal strictly
to zero number, thus triggering an incorrect execution path.

+++ core/jsMain/src/LocalDate.kt
@@ -49,7 +49,7 @@ private fun LocalDate.plusNumber(value: Number, unit: CalendarUnit): LocalDate =
         }.let(::LocalDate)
 
 public actual fun LocalDate.plus(value: Long, unit: CalendarUnit): LocalDate =
-        plusNumber(value, unit)
+        plusNumber(value.toDouble(), unit)
 
 public actual fun LocalDate.plus(value: Int, unit: CalendarUnit): LocalDate =
         plusNumber(value, unit)
